+++ hoseru/src/app/layout.tsx
@@ -7,40 +7,7 @@ import { WetherForecastThisWeek } from "../components/WetherForecastThisWeek";
 import { WetherForecastThreeDays } from "../components/WetherForecastThreeDays";
 import Link from "next/link";
 import "./globals.css";
-import { useEffect } from "react";
-import { useState } from "react";
-
-const CommanDescription = () => {
-  const urlLivedoor = 'https://weather.tsukumijima.net/api/forecast/city/220040';
-
-  interface Description{
-    bodyText:string;
-  }
-
-  const [forecastDataLivedoor,setForecastDataLivedoor] = useState(null);
-  const [detailText,setDetailText]= useState<Description>({ bodyText:"天気の概要"});
-
-  useEffect(() => {
-      const fetchData = async () => {
-          try {
-              const responseLivedoor = await fetch(urlLivedoor);
-              const dataLivedoor = await responseLivedoor.json();
-              setForecastDataLivedoor(dataLivedoor.forecasts);
-              setDetailText({bodyText: dataLivedoor.description.bodyText || ""});
-          } catch (error) {
-              console.error('Error:', error);
-          }
-      };
-
-      fetchData();
-  }, [urlLivedoor]);
-
-  return (
-    <div>
-      <p dangerouslySetInnerHTML={{ __html: detailText.bodyText ? detailText.bodyText.replace(/\n+/g, '<br />') : "" }}></p>
-    </div>
-  );
-};
+import { CommonDescription } from "../components/CommonDescription";
 
 
 export default function Layout({children}:{children: React.ReactNode}) {
@@ -86,7 +53,7 @@ export default function Layout({children}:{children: React.ReactNode}) {
           </div>
 
 
-          < CommanDescription/ >
+          < CommonDescription/ >
 
           <div>
             {children}
